% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{backward_pipes}
\alias{backward_pipes}
\title{Apply Backward Transformations to Multiple Pipes}
\usage{
backward_pipes(pipes, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipes}{List of parsed pipes from \code{parse_pipes()}}

\item{env}{Environment where the pipes will be evaluated}

\item{enclos}{Enclosing environment for evaluation of the closures}
}
\value{
List of results from applying backward transformations to each pipe
}
\description{
Simple wrapper for applying \code{backward_single_pipe()} over multiple
pipes. This enables inverse transformations for multiple reversible
pipelines.
}
\examples{
\dontrun{
library(data.table)
data(simple_dt_date)

gen_log_exp <- function(...) {
    list(
        forward = function(x) log(abs(x) + 1),
        backward = function(x) exp(x) - 1
    )
}

data_list <- list(data = simple_dt_date$X1)
raw_pipes <- list(
    list(on = "data", transforms = list(list(fun = "gen_log_exp")))
)

parsed <- parse_pipes(raw_pipes, env = data_list)
forward_results <- forward_pipes(parsed, env = data_list)
backward_results <- backward_pipes(
    parsed,
    env = list(data = forward_results[[1]])
)
}

}
\seealso{
\itemize{
\item \code{\link{backward_single_pipe}} for single pipe evaluation
\item \code{\link{forward_pipes}} for applying forward transformations
\item \code{\link{parse_pipes}} for creating multiple parsed pipes
\item \code{\link{combine_pipes}} for combining evaluation results
}
}
