% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{combine_features}
\alias{combine_features}
\title{Combines Slice Two Features Into One}
\usage{
combine_features(slice, feature1, feature2, return.all = FALSE)
}
\arguments{
\item{slice}{a \code{slice_artifact} object}

\item{feature1, feature2}{the features which will be combined into one}

\item{return.all}{boolean indicating if the returned object should contain all remaining features
along the merged one}
}
\description{
Concatenates two features of a slice into a single one
}
