% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{dim.slice_artifact}
\alias{dim.slice_artifact}
\title{Get Dimensions Of \code{slice_artifact} Objects}
\usage{
\method{dim}{slice_artifact}(x)
}
\arguments{
\item{x}{a \code{slice_artifact} object}
}
\description{
Returns a vector with number of features and time indexes, in this order
}
