% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{eval_pipes}
\alias{eval_pipes}
\title{Avalia Lista De Pipes}
\usage{
eval_pipes(pipes, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipes}{lista de pipes ja parsed por \code{parse_pipes}}

\item{env}{ambiente onde o pipe sera avaliado}

\item{enclos}{ambiente de encerramento para avaliacao das closures}
}
\value{
data.table unico combinando os resultados da aplicacao de todos os pipes em \code{pipes}
}
\description{
Wrapper simples para loop de \code{eval_single_pipe} em multiplos pipes
}
