% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{eval_single_pipe}
\alias{eval_single_pipe}
\title{Avalia Um Unico Pipe}
\usage{
eval_single_pipe(pipe, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipe}{um pipe parsed por \code{parse_single_pipe}}

\item{env}{ambiente onde o pipe sera avaliado}

\item{enclos}{ambiente de encerramento para avaliacao das closures}
}
\value{
resultado da aplicacao das closures em \code{transforms} ao(s) dado(s) em \code{"on"}
}
\description{
Aplica as closures em um \code{pipe} já parsed ao dado definido em \code{"on"}
}
\details{
Ambos \code{env} e \code{enclos} funcionam tal qual como em \code{eval}, permitindo que \code{env} corresponda a uma
lista nomeada ou data.frame, e \code{enclos} seja um ambiente onde funcoes auxiliares podem ser
definidas
}
