% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_datasets}
\alias{example_datasets}
\alias{simple_dt_date}
\alias{simple_dt_datetime}
\alias{keyed_dt_date}
\alias{keyed_dt_datetime}
\title{Example Datasets}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 5 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 5 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 100 rows and 6 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 100 rows and 6 columns.
}
\usage{
simple_dt_date

simple_dt_datetime

keyed_dt_date

keyed_dt_datetime
}
\description{
Sample \code{data.table}s for testing and example code in this package
}
\details{
There are four datasets included:

\itemize{
\item simple_dt_date
\item simple_dt_datetime
\item keyed_dt_date
\item keyed_dt_datetime
}

In the following, all values are merely illustrative and have no bearing on the actual datasets.

The ones prefixed \code{simple} have a single column of time index, which can be of either type
\code{Date} or \code{POSIXct}, as indicaded by the suffixes. \code{simple_dt_date} looks like\tabular{lllll}{
   date \tab X1 \tab X2 \tab X3 \tab Y \cr
   2025-01-02 \tab 16 \tab -18 \tab 66 \tab 132 \cr
   2025-01-03 \tab 5 \tab -17 \tab 55 \tab 122 \cr
   2025-01-04 \tab 12 \tab -16 \tab 68 \tab 138 \cr
   2025-01-05 \tab 15 \tab -19 \tab 69 \tab 137 \cr
   2025-01-06 \tab 9 \tab -7 \tab 70 \tab 125 \cr
}


\code{simple_dt_datetime} is similar, except instead of column \code{date}, there is a column
named \code{datetime} of hour by hour times.

Keyed datasets are very similar, except for having an aditional time index column. For example,
\code{keyed_dt_date} looks like\tabular{llllll}{
   date \tab target_date \tab X1 \tab X2 \tab X3 \tab Y \cr
   2025-01-02 \tab 2025-01-03 \tab 16 \tab -18 \tab 66 \tab 132 \cr
   2025-01-02 \tab 2025-01-04 \tab 5 \tab -17 \tab 55 \tab 122 \cr
   2025-01-02 \tab 2025-01-05 \tab 12 \tab -16 \tab 68 \tab 138 \cr
   2025-01-02 \tab 2025-01-06 \tab 15 \tab -19 \tab 69 \tab 137 \cr
   2025-01-02 \tab 2025-01-07 \tab 9 \tab -7 \tab 70 \tab 125 \cr
}


With \code{keyed_dt_datetime} being analogous.
}
\keyword{datasets}
