% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{forward_pipes}
\alias{forward_pipes}
\title{Apply Forward Transformations to Multiple Pipes}
\usage{
forward_pipes(pipes, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipes}{List of parsed pipes from \code{parse_pipes()}}

\item{env}{Environment where the pipes will be evaluated}

\item{enclos}{Enclosing environment for evaluation of the closures}
}
\value{
List of results from applying forward transformations to each pipe
}
\description{
Simple wrapper for applying \code{forward_single_pipe()} over multiple
pipes. This is the recommended way to apply forward transformations to
multiple pipes.
}
\examples{
\dontrun{
library(data.table)
data(simple_dt_date)

gen_standardize <- function(cols, x, ...) {
    train_means <- colMeans(x[, ..cols], na.rm = TRUE)
    train_sds <- apply(x[, ..cols], 2, sd, na.rm = TRUE)
    function(x) {
        x_copy <- copy(x)
        for (col in cols) {
            x_copy[[col]] <- (x_copy[[col]] - train_means[col]) / train_sds[col]
        }
        return(x_copy)
    }
}

data_list <- list(data = simple_dt_date)
raw_pipes <- list(
    list(on = "data", transforms = list(list(fun = "gen_standardize", cols = "X1"))),
    list(on = "data", transforms = list(list(fun = "gen_standardize", cols = "X2")))
)

parsed <- parse_pipes(raw_pipes, env = data_list)
results <- forward_pipes(parsed, env = data_list)
}

}
\seealso{
\itemize{
\item \code{\link{forward_single_pipe}} for single pipe evaluation
\item \code{\link{backward_pipes}} for applying inverse transformations
\item \code{\link{parse_pipes}} for creating multiple parsed pipes
\item \code{\link{combine_pipes}} for combining evaluation results
\item \code{\link{eval_pipes}} (deprecated) for legacy evaluation
}
}
