% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-methods.r
\name{general-methods}
\alias{general-methods}
\alias{mean.slice_artifact}
\alias{min.slice_artifact}
\alias{max.slice_artifact}
\alias{head.slice_artifact}
\alias{tail.slice_artifact}
\alias{median.slice_artifact}
\alias{summary.slice_artifact}
\alias{sd.slice_artifact}
\alias{var.slice_artifact}
\alias{quantile.slice_artifact}
\title{General Methods For \code{slice_artifact} Class}
\usage{
\method{mean}{slice_artifact}(x, variables, ...)

\method{min}{slice_artifact}(x, variables, ...)

\method{max}{slice_artifact}(x, variables, ...)

\method{head}{slice_artifact}(x, variables, ...)

\method{tail}{slice_artifact}(x, variables, ...)

\method{median}{slice_artifact}(x, na.rm = FALSE, variables, ...)

\method{summary}{slice_artifact}(object, variables, ...)

\method{sd}{slice_artifact}(x, variables, ...)

\method{var}{slice_artifact}(x, variables, ...)

\method{quantile}{slice_artifact}(x, variables, ...)
}
\arguments{
\item{x}{a \code{slice_artifact} object}

\item{variables}{variables on which to compute function}

\item{...}{remaining arguments passed on to generic}

\item{na.rm}{boolean indicating if \code{NA}s should be ignored}

\item{object}{a \code{slice_artifact} object}
}
\value{
\code{slice_artifact} with computed function on all slices of specified \code{variables}
}
\description{
Methods of \link{base}, \link{utils} and \link{stats} generics for \code{slice_artifact} objects
}
