% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-utils.r
\name{get_cluster}
\alias{get_cluster}
\title{Gets Internal \code{shapeshiftr} Cluster}
\usage{
get_cluster()
}
\value{
object of type \code{cluster}
}
\description{
Internal function for accessing package dedicated cluster
}
