% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{na.exclude.slice_artifact}
\alias{na.exclude.slice_artifact}
\title{Remove \code{NA} From Slices}
\usage{
\method{na.exclude}{slice_artifact}(object, ...)
}
\description{
Removes all indexes where there is at least one feature with \code{NA}s
}
