% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{names_slice_artifact}
\alias{names_slice_artifact}
\alias{names.slice_artifact}
\alias{names<-.slice_artifact}
\title{Get And Set Names Of \code{slice_artifact} Object}
\usage{
\method{names}{slice_artifact}(x)

\method{names}{slice_artifact}(x) <- value
}
\description{
Functions for getting names of variables and renaming them in an \code{slice_artifact} object
}
