% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeshiftr_closure.r
\name{new_shapeshiftr_closure}
\alias{new_shapeshiftr_closure}
\title{Internal \code{shapeshiftr_closure} Constructor}
\usage{
new_shapeshiftr_closure(funs)
}
\arguments{
\item{funs}{Either single function or list of two functions. If single function, it is assumed
to be the forward transformation and backward is set to identity. If list, assumed to be
forward and backward in this order}
}
\value{
S3 object of class \code{shapeshiftr_closure} with elements:
\item{forward}{Forward transformation function}
\item{backward}{Backward transformation function}
}
\description{
S3 wrapper for paired forward/backward closures enabling reversible transformations
}
\details{
If \code{funs} is a single function, it becomes the forward transformation
and \code{function(x) x} is used as the backward transformation (identity).
}
\keyword{internal}
