% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{parse_pipes}
\alias{parse_pipes}
\title{Parse Lista De Pipes}
\usage{
parse_pipes(raw_pipes)
}
\arguments{
\item{raw_pipes}{lista definindo diversos pipes}
}
\value{
lista \code{raw_pipes} com elementos \code{"transforms"} de cada pipe avaliados
}
\description{
Wrapper simples para loop de \code{parse_single_pipe} em multiplos pipes
}
