% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{parse_single_pipe}
\alias{parse_single_pipe}
\title{Parse a Single Pipe}
\usage{
parse_single_pipe(raw_pipe, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{raw_pipe}{a list defining a singular pipe, that is, with element \code{"on"} and \code{"transforms"}}

\item{env}{environment where the raw pipes will be evaluated}

\item{enclos}{enclosing environment for evaluation of the closures}
}
\value{
list \code{raw_pipe} with element \code{"transforms"} evaluated to the defined closures
}
\description{
Interprets the definition of a single pipe, generating the defined closures
}
\details{
Both \code{env} and \code{enclos} function just as in \code{eval}, allowing \code{env} to correspond to a named
list or data.frame, and \code{enclos} to be an environment where helper functions can be defined

\code{raw_pipe} should be a named list with two arguments: \code{"on"} and \code{"transforms"}. \code{"on"} is
simply a vector of at most two elements indicating the names of the variables on which
the transformation(s) operate(s). \code{"transforms"} should be a list of transformation specifications.
Each transformation is in turn a list with at least one named element, \code{"fun"}, a string indicating
the name of the closure generator to be called. In addition to \code{"fun"}, other elements may exist
indicating parameterization of the closure generator. See \code{Examples} for an illustration of a raw pipe.

Note that functions in transformation definitions must always be closure generators and must
necessarily include the variadic argument \code{...}. This is necessary because, at a later stage,
these generators will be interpreted to then produce the parsed pipe in which \code{transforms} are,
in fact, functions to be called on the data.
}
\examples{

# generator of a closure with single argument `x` that filters this argument based on
# other specified parameters
gen_closure_filter <- function(by, value, ...) function(x) x[x[[by]] == value, ]

# generator of a closure with single argument `x` that calls `summary` on this argument
gen_closure_summary <- function(...) function(x) summary(x)

raw_pipe <- list(
    on = "mtcars",
    transforms = list(
        list(
            fun = "gen_closure_filter",
            by = "cyl", value = 6
        ),
        list(
            fun = "gen_closure_summary"
        )
    )
)

parsed_pipe <- shapeshiftr:::parse_single_pipe(raw_pipe)

}
