% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{refresh_cluster}
\alias{refresh_cluster}
\alias{refresh_cluster.default}
\alias{refresh_cluster.cluster}
\title{Updates Cluster With Contextual Information}
\usage{
refresh_cluster(cl, envir)

\method{refresh_cluster}{default}(cl, envir)

\method{refresh_cluster}{cluster}(cl, envir)
}
\arguments{
\item{cl}{an object of type \code{cluster}}

\item{envir}{environment from which to export names}
}
\value{
nothing, only has side effect of exporting names in \code{envir} to \code{cl}
}
\description{
Exports names in current calling environment to cluster \code{cl}
}
\details{
Given that the cluster is initialized when the package is loaded, it needs to receive, at each
parallel function execution, the objects necessary for computation. \code{refresh_cluster} looks
for all names in an environment \code{envir} and exports them to cluster \code{cl}

Since this is an internal function, it is not meant to be directly called by the user. Normally
it is seen in functions as \code{refresh_cluster(cl, parent.frame())}, i.e., refreshing with the
calling environment.
}
