% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{run_loop}
\alias{run_loop}
\title{Internal Loop Executor}
\usage{
run_loop(X, fun)
}
\arguments{
\item{X}{iterable collection}

\item{function}{to be applied}
}
\value{
list with results of applying \code{fun} to each element of \code{X}
}
\description{
Simple wrapper to run loops, either single or multithreaded
}
