% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-utils.r
\name{run_post_hook}
\alias{run_post_hook}
\title{Treatment Of Cluster On Package Unload}
\usage{
run_post_hook(cl)
}
\arguments{
\item{cl}{cluster or NULL}
}
\value{
no explicit return, only stops \code{cl} if it is a cluster, or does nothing otherwise
}
\description{
If \code{cl} is a cluster, stops it
}
