% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-methods.r
\name{sd}
\alias{sd}
\alias{sd.default}
\title{Mock Generic For Standard Deviation}
\usage{
sd(x, ...)

\method{sd}{default}(x, ...)
}
\arguments{
\item{x}{a \code{slice_artifact} object}

\item{...}{remaining arguments of \code{\link[stats:sd]{stats::sd()}}}
}
\description{
Reexports the \code{\link[stats:sd]{stats::sd()}} function as a generic so new methods can be created
}
\seealso{
\link{general-methods}
}
