% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-utils.r
\name{set_up_cluster}
\alias{set_up_cluster}
\title{Configures Internal Package Cluster}
\usage{
set_up_cluster(threads)
}
\arguments{
\item{threads}{number of cluster threads}
}
\value{
has no explicit return. Assings a FORK cluster to name \code{.SHAPESHIFTR_CLUSTER} in the
package's namespace, if code{threads > 1}, or a NULL object otherwise
}
\description{
Sets up a FORK cluster of \code{threads} threads, if running on linux platforms
}
\details{
This is an internal function, should not be called directly by the user
}
