% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeshiftr.r
\docType{package}
\name{shapeshiftr}
\alias{shapeshiftr}
\title{Package \code{shapeshiftr}}
\description{
Automated pipeline for shape manipulation and data transforms
}
\details{
This package aims at providing a simple and efficient interface for manipulation and
transformation of data for training forecasting models. This is effectively done through
"slicing" operations, which transform standard tabular time series data into data.frames usable
for model fitting/forecasting, as well as diverse transforms.
}
