% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_population.R
\docType{data}
\name{split_population}
\alias{split_population}
\title{Data to showcase a population fragmenting into two subpopulations}
\format{
A data frame with 100 rows and 5 columns.
\describe{
\item{x}{Longitude of animal}
\item{y}{Latitude of animal}
\item{year}{Year animal was sampled}
\item{random}{A column containing data but not relevant to these functions}
\item{id}{The name of the animal sampled}
}
}
\source{
Created by Nicola Love, by creating 3 bivariate normal distributions using \code{MASS::mvrnorm()}. All three distributions had a standard deviation of 15, the distribution in year 1 was centered at c(50,50), the two distributions in year 2 were centered at c(50,0) and c(50, 100).
}
\usage{
split_population
}
\description{
A dataset containing locations of sampled animals over 2 years
}
\examples{
data(split_population)
}
\keyword{datasets}
