% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-methods.r
\name{var}
\alias{var}
\alias{var.default}
\title{Mock Generic For Variance}
\usage{
var(x, ...)

\method{var}{default}(x, ...)
}
\arguments{
\item{x}{a \code{slice_artifact} object}

\item{...}{remaining arguments of \code{\link[stats:cor]{stats::var()}}}
}
\description{
Reexports the \code{\link[stats:cor]{stats::var()}} function as a generic so new methods can be created
}
\seealso{
\link{general-methods}
}
