% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{as.data.table.slice_artifact}
\alias{as.data.table.slice_artifact}
\title{Converts \code{slice_artifact} To A \code{data.table}}
\usage{
\method{as.data.table}{slice_artifact}(x, ...)
}
\description{
Transforms an object of class \code{slice_artifact} to a \code{data.table}
}
