% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{combine_pipes}
\alias{combine_pipes}
\title{Combination of Pipe Results}
\usage{
combine_pipes(evals, combine_fun = default_combine, ...)
}
\arguments{
\item{evals}{list of elements to be combined, typically the result of \code{eval_pipes}}

\item{combine_fun}{combination function to be applied, by default \code{default_combine}. See
Details}

\item{...}{additional arguments to be passed to \code{combine_fun}}
}
\value{
result of the combination
}
\description{
Combines outputs from pipes resulting from a call to \code{eval_pipes}
}
\details{
Essentially this function is a wrapper of \code{Reduce} applied to \code{evals} with the function
\code{combine_fun}. Because of this, \code{combine_fun} should be a function with at least two arguments
\code{x} and \code{y}, as well as the variadic argument \code{...} for consistency between calls. Other
optional arguments are permitted and will be passed to \code{combine_fun}.

\code{default_combine} is merely a call to \code{merge} using \code{by = 1}, that is, merging all
data in \code{evals} by the first column of each one.
}
