% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{combine_pipes}
\alias{combine_pipes}
\title{Combinacao De Resultados De Pipes}
\usage{
combine_pipes(evals, combine_fun = default_combine, ...)
}
\arguments{
\item{evals}{lista de elementos a serem combinados, tipicamente o resultado de \code{eval_pipes}}

\item{combine_fun}{funcao de combinacao a ser aplicada, por default \code{default_combine}. Veja
Detalhes}

\item{...}{argumentos adicionais a serem passados para \code{combine_fun}}
}
\value{
resultado da combinacao
}
\description{
Combina saidas de pipes resultantes de uma chamada de \code{eval_pipes}
}
\details{
Essencialmente esta funcao e um wrapper de \code{Reduce} aplicado a \code{evals} com a funcao
\code{combine_fun}. Em funcao disto, \code{combine_fun} deve ser uma funcao de minimamente dois argumentos
\code{x} e \code{y}, bem como o argumento variatico \code{...} para consistencia entre chamadas. Demais
argumentos opcionais sao permitidos e serao repassados para \code{combine_fun}.

\code{default_combine} e apenas uma chamada de \code{merge} utilizando \code{by = 1}, isto e, merge de todos
os dados em \code{evals} pela primeira coluna de cada um deles.
}
