% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvpop.R
\name{cvpop}
\alias{cvpop}
\title{Calculate Coefficient of Variation (CV) of Inter-individual Distances at the Population-level}
\usage{
cvpop(points, log10 = FALSE)
}
\arguments{
\item{points}{A data frame, matrix, or vector of locations, where each row is a sampled individual.}

\item{log10}{An optional argument for the log transformation of the CV of inter-individual distances.}
}
\value{
A single numeric value for calculated coefficient of variation of the inter-individual distances between every pair of individuals. If this is used is a tidyverse pipe, the output will be a column of numbers.
}
\description{
This function calculates the CV of the inter-individual distance between every pair of individuals, optionally applying a log10 transformation. We assume all individuals are sampled approximately within the same time point. If the input is a data frame of multiple sampling sessions, we suggest using this function in a pipe where the data frame is grouped, e.g., data \%>\% group_by(day, year) \%>\% summarise(cvpop = cvpop(c(x,y)))
}
