% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvratio.R
\name{cvratio}
\alias{cvratio}
\title{Calculate the Ratio of Individual- to Population-level Coefficient of Variation (CV) of Inter-individual Distances}
\usage{
cvratio(points, log10 = FALSE)
}
\arguments{
\item{points}{A data frame, matrix, or vector of locations, where each row is a sampled individual.}

\item{log10}{An optional argument for the log transformation of the ratio of CVs.}
}
\description{
This function calculates the ratio of individual-level CVs to population-level CVs of inter-individual distances, optionally applying a log10 transformation. We assume all individuals are sampled approximately within the same time point. If the input is a data frame of multiple sampling sessions, we suggest using this function in a pipe where the data frame is grouped, e.g., data \%>\% group_by(day, year) \%>\% summarise(cvratio = cvratio(c(x,y)))
}
