% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.r
\name{dwt.slice_artifact}
\alias{dwt.slice_artifact}
\alias{dwt.default}
\alias{dwt}
\title{Discrete Wavelet Transform Method For \code{slice_artifact} Objects}
\usage{
\method{dwt}{slice_artifact}(X, variables, ...)

\method{dwt}{default}(X, ...)

dwt(X, ...)
}
\arguments{
\item{X}{a \code{slice_artifact} object}

\item{variables}{variables on which to perform the discrete wavelet transform}

\item{...}{remaining arguments of \code{\link[wavelets]{dwt}}}
}
\description{
Extends the \code{\link[wavelets]{dwt}} function to operate on \code{slice_artifact}
}
