% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{eval_pipes}
\alias{eval_pipes}
\title{Internal Helper for Evaluating Multiple Pipes}
\usage{
eval_pipes(pipes, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipes}{List of pipes with raw function closures}

\item{env}{Environment for evaluation}

\item{enclos}{Enclosing environment for closures}
}
\value{
List of transformed data
}
\description{
Internal wrapper that loops \code{eval_single_pipe()} to multiple pipes, not intended for direct use
}
\seealso{
\itemize{
\item \code{\link{forward_pipes}} for applying forward transformations
\item \code{\link{backward_pipes}} for applying backward transformations
}
}
\keyword{internal}
