% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{eval_pipes}
\alias{eval_pipes}
\title{Avalia Lista De Pipes}
\usage{
eval_pipes(pipes)
}
\arguments{
\item{pipes}{lista de pipes ja parsed por \code{parse_pipes}}
}
\value{
data.table unico combinando os resultados da aplicacao de todos os pipes em \code{pipes}
}
\description{
Wrapper simples para loop de \code{eval_single_pipe} em multiplos pipes
}
