% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{eval_pipes}
\alias{eval_pipes}
\title{Evaluate List of Pipes}
\usage{
eval_pipes(pipes, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipes}{list of pipes already parsed by \code{parse_pipes}}

\item{env}{environment where the pipe will be evaluated}

\item{enclos}{enclosing environment for evaluation of the closures}
}
\value{
single data.table combining the results of applying all pipes in \code{pipes}
}
\description{
Simple wrapper for looping \code{eval_single_pipe} over multiple pipes
}
\seealso{
\itemize{
\item \code{\link{eval_single_pipe}} for details on parse-eval separation
\item \code{\link{parse_pipes}} for creating multiple parsed pipes
}
}
