% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{eval_single_pipe}
\alias{eval_single_pipe}
\title{Evaluate a Single Pipe}
\usage{
eval_single_pipe(pipe, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{pipe}{a pipe parsed by \code{parse_single_pipe}}

\item{env}{environment where the pipe will be evaluated}

\item{enclos}{enclosing environment for evaluation of the closures}
}
\value{
result of applying the closures in \code{transforms} to the data in \code{"on"}
}
\description{
Applies the closures in a \code{pipe} already parsed to the data defined in \code{"on"}
}
\details{
Both \code{env} and \code{enclos} function just as in \code{eval}, allowing \code{env} to correspond to a named
list or data.frame, and \code{enclos} to be an environment where helper functions can be defined
}
