% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{merge.slice_artifact}
\alias{merge.slice_artifact}
\title{Merges \code{slice_artifact} Objects}
\usage{
\method{merge}{slice_artifact}(x, y, ...)
}
\arguments{
\item{x, y}{objects of class \code{slice_artifact} for merging}

\item{...}{has no utility other than consistency with the generic}
}
\description{
Internal function; merges objectes by common index into a single \code{slice_artifact}
}
\details{
This function is not intended to be directly called by the user. It is only deployed in specific
conditions within the package and does not have all the necessary sanity checks for external use.

Contrary to \code{c.slice_artifact}, this function is meant to combine artifacts from different
data, or different slicing parameters, into a unified artifact object. This will match indexes
in all elements for merging and ONLY USE indexes which are common to all. There is no \code{all}
argument like in data.frame merging.
}
