% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{new_slice_artifact}
\alias{new_slice_artifact}
\title{Slice Artifact Inner Constructor}
\usage{
new_slice_artifact(list, index, L)
}
\arguments{
\item{list}{a list of data.frame-like slices as extracted by \code{do_single_slice}}

\item{index}{scalar of vector of time indexes corresponding to each slice}

\item{L}{the list of lag/leads used in extracting slices}
}
\description{
Builds an object of class \code{slice_artifact}
}
