% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{new_slice_artifact}
\alias{new_slice_artifact}
\title{Slice Artifact Inner Constructor}
\usage{
new_slice_artifact(list, index, L, H)
}
\description{
Builds an object of class \code{slice_artifact}
}
