% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{parse_pipes}
\alias{parse_pipes}
\title{Parse List of Pipes}
\usage{
parse_pipes(raw_pipes, env = parent.frame(), enclos = parent.frame())
}
\arguments{
\item{raw_pipes}{list defining various pipes}

\item{env}{environment where the pipe will be evaluated}

\item{enclos}{enclosing environment for evaluation of the closures}
}
\value{
list \code{raw_pipes} with elements \code{"transforms"} of each pipe evaluated
}
\description{
Simple wrapper for looping \code{parse_single_pipe} over multiple pipes
}
