% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.r
\name{parse_single_pipe}
\alias{parse_single_pipe}
\title{Parse Um Unico Pipe}
\usage{
parse_single_pipe(raw_pipe, env, enclos = parent.frame())
}
\arguments{
\item{raw_pipe}{uma lista definindo pipe singular, isto e, com elemento \code{"on"} e \code{"transforms"}}

\item{env}{ambiente onde os pipes brutos serao avaliados}

\item{enclos}{ambiente de encerramento para avaliacao das closures}
}
\value{
lista \code{raw_pipe} com elemento \code{"transforms"} avaliado para as closures definidas
}
\description{
Interpreta a definicao de um unico pipe, gerando as closures definidas
}
\details{
Ambos \code{env} e \code{enclos} funcionam tal qual como em \code{eval}, permitindo que \code{env} corresponda a uma
lista nomeada ou data.frame, e \code{enclos} seja um ambiente onde funcoes auxiliares podem ser
definidas
}
