% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeshiftr.r
\docType{package}
\name{shapeshiftR}
\alias{shapeshiftR}
\title{Package \code{shapeshiftr}}
\description{
Automated pipeline for shape manipulation and data transforms
}
\details{
This package intends to unify several kinds of data wrangling under a single framework.
This includes shape manipulation as well data transformations. All is meant to be easily
readable and extensible, aiming at greatly simplifying the process of data preparation before
fitting/using models.
}
