% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeshiftr-package.R
\docType{package}
\name{shapeshiftr}
\alias{shapeshiftr-package}
\alias{shapeshiftr}
\title{shapeshiftr: Analyze Group/Home Range Shape Shifts Using Location Data.}
\description{
R package for detecting population range reconfiguration via inter-individual distances
}
\details{
The package is designed for spatial and temporal analysis of location data, especially those multiple sampling sessions (seasons, years, sites, populations, etc).
\subsection{Reference}{

For the full description of the method and its application, see:
Love & Otto (2025). \emph{Spatial Signatures of Population Cohesion}. Methods in Ecology and Evolution. https://doi.org/##
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nicolalove.github.io/shapeshiftr/}
}

}
\author{
\strong{Maintainer}: Nicola Love \email{nicolalove@zoology.ubc.ca}

Authors:
\itemize{
  \item Sarah Otto
}

}
