% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_artifact.r
\name{[.slice_artifact}
\alias{[.slice_artifact}
\title{Subsetting Of \code{slice_artifact}}
\usage{
\method{[}{slice_artifact}(x, i, j, ...)
}
\arguments{
\item{x}{object of class \code{slice_artifact}}

\item{i}{variables for subsetting; if missing returns all variables}

\item{j}{indexes for subsetting; if missing returns all indexes}
}
\description{
Internal function; subsets \code{slice_artifact} either by variable or index
}
\details{
This function is not intended to be directly called by the user. It is only deployed in specific
conditions within the package and does not have all the necessary sanity checks for external use.
}
